jira_issue_tracker/requirements.txt
jira
pandas
openpyxl
python-dotenv
urllib3